% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_proj.R
\name{exp_proj}
\alias{exp_proj}
\title{Exponential Growth Projection}
\usage{
exp_proj(N_0, r, t)
}
\arguments{
\item{N_0}{starting population size value}

\item{r}{per capita growth rate value}

\item{t}{time at which to project}
}
\value{

}
\description{
Exponential Growth Projection
}
\examples{


}
